Add image grid UI

+++ app/src/main/kotlin/com/jzbrooks/splashdown/data/ImageDataSource.kt
@@ -1,6 +1,6 @@
 package com.jzbrooks.splashdown.data
 
 interface ImageDataSource {
-    suspend fun getRecentPhotos(page: Long): PhotoResult
-    suspend fun searchPhotos(query: String, page: Long): PhotoResult
+    suspend fun getRecentPhotos(page: Int): PhotoResult
+    suspend fun searchPhotos(query: String, page: Int): PhotoResult
 }
