+++ app/src/main/kotlin/com/jzbrooks/splashdown/data/network/FlickerService.kt
@@ -6,8 +6,8 @@ import retrofit2.http.Query
 
 interface FlickerService {
     @GET("?method=flickr.photos.getRecent&api_key=${BuildConfig.flickrApiKey}&format=json&nojsoncallback=1")
-    suspend fun getRecentPhotos(@Query("page") page: Long): PhotosResponseDto
+    suspend fun getRecentPhotos(@Query("page") page: Int): PhotosResponseDto
 
     @GET("?method=flickr.photos.search&api_key=${BuildConfig.flickrApiKey}&format=json&nojsoncallback=1")
-    suspend fun searchPhotos(@Query("text") query: String, @Query("page") page: Long): PhotosResponseDto
+    suspend fun searchPhotos(@Query("text") query: String, @Query("page") page: Int): PhotosResponseDto
 }
